📁 Rename : types, methods 파일 별도 분리
